feat: Add Block Storage API (#151)
This starts adding the block storage API alias cinder. So far it implements rudimentary versions of list_volumes, find_volumes and get_volume with an equally minimal Volume struct. But I does compile and returns things.

Although the cinder API is very similar to nova's API, this is currently closer to the image module. For example, at the moment it goes directly to the /volumes/detail endpoint, rather than defining some kind of VolumeSummary for /volumes and offering a detailed() as the compute module does.

This is for simplicity, just to get things rolling. I need this at work, so I can invest a fair bit of time. I'll make it more consistent along the way.